move validate slack validation
